backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]